signed tile addressing bug fix
It wasn't converting to a 16 bit signed offset correctly, super mario land looks normal which means this is now working correctly.

+++ mem.c
@@ -400,7 +400,7 @@ void WB(struct GB* gb, uint16_t addr, uint8_t val, uint8_t cycles)
 const struct tileStruct* tile_access(struct GB* gb, uint16_t map_base, uint16_t map_index)
 {	// Obtain the tile index from the tile map, selected by the map_base. Important, map_base
 	//		should only ever be 0x9800 or 0x9C00 as there are only two tile maps
-	uint16_t tiledata_index = gb->vram[map_base + map_index - VRAM_BASE_ADDR];
+	uint8_t tiledata_index = gb->vram[map_base + map_index - VRAM_BASE_ADDR];
 	// Tiles are indexed one of two ways, using either signed or unsigned addressing. This is
 	//		determined by LCDC bit 4
 	if (gb->ppu.bgwin_tiledata)
@@ -409,7 +409,7 @@ const struct tileStruct* tile_access(struct GB* gb, uint16_t map_base, uint16_t 
 	}
 	else
 	{	// Signed tile addressing, uses 0x9000 as base pointer
-		int16_t signed_index = (int16_t)tiledata_index;
+		int16_t signed_index = (int16_t)(int8_t)tiledata_index;
 		return (struct tileStruct*)&gb->vram[0x1000+(signed_index*sizeof(struct tileStruct))];
 	}
 }
